Refactor enroll secrets components into own directory
