+++ frontend/components/EnrollSecrets/DeleteSecretModal/_styles.scss
