+++ frontend/components/EnrollSecrets/DeleteSecretModal/index.ts
