+++ frontend/components/EnrollSecrets/EnrollSecretTable/_styles.scss
