+++ frontend/components/EnrollSecrets/SecretEditorModal/_styles.scss
