+++ frontend/components/EnrollSecrets/SecretEditorModal/index.ts
